project/requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.2
django-prometheus==2.3.1
gunicorn==23.0.0
whitenoise==6.8.2